+++ .vscode/settings.json
@@ -4,6 +4,7 @@
         "Callouts",
         "Codeflow",
         "Hacky",
+        "Rebaser",
         "Resizer",
         "Retval",
         "Sarif",
