+++ src/extension/uriRebaser.ts
